v1/requirements.txt
numpy==1.16.4
matplotlib==3.1.0
pyzmq==18.0.0
scikit-learn==0.21.2
scikit-image==0.15.0
scipy==1.3.0

